Properties: Rename to BasicProperties, and modularise for different Node types
